Added funcitonality to filter panel [#158037410]
Batch and Location drop downs are now populated with data dynamically.
This data is taken from batch store for batches and a set of locations from rooms in the room store.
Added a checkbox for hasBedAvailable or hide full rooms.
Batch drop down now assigns batch id to the filter when updating.